+++ vaadin-form-layout.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-form-layout.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
